run docketsync from within a Docker container
Including a Dockerfile and Makefile for building Docker images with the
production and developement Python3 packaged needed to run `docketsync`.

Build the Docker image with:

```
make build
```

Run `docketsync` inside of a Docker container with:

```
./docketsync-run [options]
```

where `[options]` are options to the `docketsync` command. To see all of
the options, use the command:

```
./docketsync-run --help
```

For example, the following command will download comments from docket id
`USCIS-2010-0012` that were posted to the website from Dec 05, 2018
through Dec 07, 2018 and save the comment metadata and text to a sqlite
database and csv file:

```
./docketsync-run \
  --docket USCIS-2010-0012 \
  --posted-date 12/05/18-12/07/18 \
  --save-db comments_2018-12-05_2018-12-07.db \
  --save-csv comments_2018-12-05_2018-12-07.csv
```

Users can also get into the Docker container developement environment
through the `shell` target in the Makefile. The command

```
make shell
```

launches the Docker container and places the user in a Bash shell. From
the shell, the user can launch an IPython3 interpreter by issuing the
command `ipython3` or can launch the `dockersync` command.

Fixing the --posted-date flag in docketsync
Adding development packages to Pipfile.